backend/requirements.txt
fastapi
uvicorn
firebase-admin
google-cloud-bigquery
pydantic
dotenv